+++ src/components/main.ts
@@ -1,6 +1,6 @@
 import Vue from "vue";
 import Component from "vue-class-component";
-import { mapState } from "vuex";
+import { mapGetters, mapState } from "vuex";
 
 import ChangelogModal from "./modals/changelog-modal.vue";
 import CharacterAddModal from "./modals/character-add-modal.vue";
@@ -10,12 +10,17 @@ import CharacterAddModal from "./modals/character-add-modal.vue";
     "changelog-modal": ChangelogModal,
     "character-add-modal": CharacterAddModal
   },
-  computed: mapState(["activeRequests"])
+  computed: {
+    ...mapState(["activeRequests"]),
+    ...mapGetters({ showChangelog: "isNewVersion" })
+  }
 })
 export default class Router extends Vue {
   activeRequests!: number;
 
-  /** Whether we have finished initializing our data. */
+  /**
+   * Whether we have finished initializing our data.
+   */
   hasLoaded = false;
 
   async created() {