+++ src/shims.d.ts
@@ -1,3 +1,6 @@
+// Automatically reead from package.json through webpack's DefinePlugin
+declare const CURRENT_VERSION: string;
+
 declare module "*.vue" {
   import Vue from "vue";
   export default Vue;