+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - Allow marking a character as dead. Dead characters can not spend stars or
   claim rewards.
-- Show a changelog in the page itself.
+- Show a changelog in the page itself. This is shown automatically whenever
+  there's a new version.
 
 ## [1.0.0] - 2019-06-07